# Issue/3286: renderableOrbitalKepler - optimization of point and trail rendering