--- a/modules/space/rendering/renderableorbitalkepler.cpp
+++ b/modules/space/rendering/renderableorbitalkepler.cpp
@@ -256,14 +256,14 @@ RenderableOrbitalKepler::Appearance::Appearance()
     })
     , color(ColorInfo, glm::vec3(1.f), glm::vec3(0.f), glm::vec3(1.f))
     , trailWidth(TrailWidthInfo, 2.f, 1.f, 20.f)
-    , pointSizeExponent(PointSizeExponentInfo, 1.0f, 0.f, 25.f)
+    , pointSizeExponent(PointSizeExponentInfo, 1.0f, 0.f, 11.f)
     , renderingModes(
         RenderingModeInfo,
         properties::OptionProperty::DisplayType::Dropdown
     )
     , trailFade(TrailFadeInfo, 20.f, 0.f, 30.f)
-    , enableMaxSize(EnableMaxSizeInfo, false)
-    , maxSize(MaxSizeInfo, 1.f, 0.f, 45.f)
+    , enableMaxSize(EnableMaxSizeInfo, true)
+    , maxSize(MaxSizeInfo, 5.f, 0.f, 45.f)
     , enableOutline(EnableOutlineInfo, true)
     , outlineColor(OutlineColorInfo, glm::vec3(0.f), glm::vec3(0.f), glm::vec3(1.f))
     , outlineWidth(OutlineWidthInfo, 0.2f, 0.f, 1.f)