--- a/modules/space/shaders/debrisVizPoints_gs.glsl
+++ b/modules/space/shaders/debrisVizPoints_gs.glsl
@@ -27,8 +27,8 @@
 #include "PowerScaling/powerScalingMath.hglsl"
 
 layout(lines) in;
-flat in double currentRevolutionFraction[];
-flat in double vertexRevolutionFraction[];
+flat in float currentRevolutionFraction[];
+flat in float vertexRevolutionFraction[];
 
 uniform dmat4 modelTransform;
 uniform dmat4 viewTransform;
@@ -43,103 +43,91 @@ layout(triangle_strip, max_vertices = 4) out;
 out float projectionViewDepth;
 out vec4 viewSpace;
 out vec2 texCoord;
-flat out int skip;
-
-dvec4 dz_normalization(dvec4 dv_in) {
-  dvec4 dv_out = dv_in;
-  dv_out.z = 0;
-  return dv_out;
-}
 
 void main() {
-  skip = 0;
-  
-  double cFrac = currentRevolutionFraction[0];
-  double v0Frac = vertexRevolutionFraction[0];
-  double v1Frac = vertexRevolutionFraction[1];
+  float cFrac = currentRevolutionFraction[0];
+  float v0Frac = vertexRevolutionFraction[0];
+  float v1Frac = vertexRevolutionFraction[1];
+
+  // Only create/emit vertices for segments where the head of the trail falls within
   if (cFrac >= v0Frac && cFrac <= v1Frac) {
 
-    // Interpolate position of point
-    double dFrac = abs(cFrac - v0Frac);
-    double vFrac = abs(v1Frac - v0Frac);
-    double percentage = dFrac / vFrac;
-
-    dvec4 v0Weighted = (1.0 - percentage) * gl_in[0].gl_Position;
-    dvec4 v1Weighted = (percentage) * gl_in[1].gl_Position;
-
-    dvec4 pos = v0Weighted + v1Weighted;
+    // Interpolate position of current position of the trail head
+    float dFrac = abs(cFrac - v0Frac);
+    float vFrac = abs(v1Frac - v0Frac);
+    float percentage = dFrac / vFrac;
+    
+    vec4 v0Weighted = (1.0 - percentage) * gl_in[0].gl_Position;
+    vec4 v1Weighted = (percentage) * gl_in[1].gl_Position;
+    vec4 pos = v0Weighted + v1Weighted;
     // ==========================
 
+    // Cast to float based data types
+    mat4 modelTrans = mat4(modelTransform);
+    mat4 viewTrans = mat4(viewTransform);
+    vec3 camPosWorld = vec3(cameraPositionWorld);
+
     // Calculate current vertex position to world space
-    dvec4 vertPosWorldSpace = modelTransform * pos;
+    vec4 vertPosWorldSpace = modelTrans * pos;
 
     // Calculate new axis for plane
-    dvec3 normal = normalize(cameraPositionWorld - vertPosWorldSpace.xyz);
-    dvec3 right = normalize(cross(cameraUpWorld, normal));
-    dvec3 up = normalize(cross(normal, right));
+    vec3 normal = vec3(normalize(camPosWorld - vertPosWorldSpace.xyz));
+    vec3 right = normalize(cross(vec3(cameraUpWorld), normal));
+    vec3 up = normalize(cross(normal, right));
 
     // Calculate size of points
-    double initialSize = pow(10.0, pointSizeExponent);
+    float initialSize = pow(10.0, pointSizeExponent);
     right *= initialSize;
     up *= initialSize;
 
-    double opp = length(right);
-    double adj = length(cameraPositionWorld.xyz - vertPosWorldSpace.xyz);
+    float opp = length(right);
+    float adj = length(camPosWorld - vertPosWorldSpace.xyz);
     float angle = atan(float(opp/adj));
     float maxAngle = radians(maxSize * 0.5);
 
     // Controls the point size
     if (enableMaxSize && (angle > maxAngle) && (adj > 0.0)) {
-      double correction = (adj * tan(maxAngle)) / opp;
+      float correction = (adj * tan(maxAngle)) / opp;
       right *= correction;
       up *= correction;
     }
 
-    // Calculate and set corners of the quad
-    dvec4 p0World = vertPosWorldSpace + (dvec4(up-right, 0.0));
-    dvec4 p1World = vertPosWorldSpace + (dvec4(-right-up,0.0));
-    dvec4 p2World = vertPosWorldSpace + (dvec4(right+up, 0.0));
-    dvec4 p3World = vertPosWorldSpace + (dvec4(right-up, 0.0));
+    // Calculate and set corners of the new quad
+    vec4 p0World = vertPosWorldSpace + vec4(up-right, 0.0);
+    vec4 p1World = vertPosWorldSpace + vec4(-right-up,0.0);
+    vec4 p2World = vertPosWorldSpace + vec4(right+up, 0.0);
+    vec4 p3World = vertPosWorldSpace + vec4(right-up, 0.0);
 
-    dmat4 ViewProjectionTransform = dmat4(projectionTransform) * viewTransform;
+    mat4 ViewProjectionTransform = projectionTransform * viewTrans;
 
     // Set some additional out parameters
-    viewSpace = z_normalization(
-      vec4(projectionTransform * viewTransform * modelTransform * pos)
-    );
+    viewSpace = z_normalization(projectionTransform * viewTrans * modelTrans * pos);
     projectionViewDepth = viewSpace.w;
 
-    //left-top
-    vec4 p0Screen = vec4(dz_normalization(ViewProjectionTransform * p0World));
+    // left-top
+    vec4 p0Screen = z_normalization(ViewProjectionTransform * p0World);
     gl_Position = p0Screen;
     texCoord = vec2(0.0, 0.0);
     EmitVertex();
 
-    //left-bot
-    vec4 p1Screen = vec4(dz_normalization(ViewProjectionTransform * p1World));
+    // left-bot
+    vec4 p1Screen = z_normalization(ViewProjectionTransform * p1World);
     gl_Position = p1Screen;
     texCoord = vec2(1.0, 0.0);
     EmitVertex();
 
-    //right-top
-    vec4 p2Screen = vec4(dz_normalization(ViewProjectionTransform * p2World));
+    // right-top
+    vec4 p2Screen = z_normalization(ViewProjectionTransform * p2World);
     gl_Position = p2Screen;
     texCoord = vec2(0.0, 1.0);
     EmitVertex();
 
-    //right-bot
-    vec4 p3Screen = vec4(dz_normalization(ViewProjectionTransform * p3World));
+    // right-bot
+    vec4 p3Screen = z_normalization(ViewProjectionTransform * p3World);
     gl_Position = p3Screen;
     texCoord = vec2(1.0, 1.0);
     EmitVertex();
-    
-    // Primitive
-    EndPrimitive();
-
   } 
-  else {
-    skip = 1;
-    EmitVertex();
-    EndPrimitive();
-  }
+
+  EndPrimitive();
 }