## Trail optimization
Optimization for the trail shader was achieved by using a geometry shader to cull vertices that does not contribute to the final trail. In other words, vertices which part of the trail would be fully faded (discard by fragment shader) is now removed in geometry shader and thus less data is being processed in fragment shader.


## Point optimization
Changed nearly all calculations from double to float. Was using double as a safe-guard during development and never revisited that decision. 

## Performance statistics
Everything is compared to 0.20.0 RC1

### Trail optimization
The new implementation results in a framerate that does not fluctuate depending on how many pixels the trails cover on screen.
Test using the _Main Asteroid Belt_ at different zoom levels within the solar system

**Before:**
- 60 - 170 fps
- 16.67 - 5.88 ms

**After:**
- 156 fps
- 6.41 ms

At most it increased by 0.53 ms and at most it recued by 10.26 ms


### Point optimization
Test using the _Main Asteroid Belt_ at different zoom levels within the solar system

**Before:**
-  86 fps
- 11.63 ms

**After:**
- 167 - 176 fps
- 5.99 - 5.68 ms

At most it reduced by 9.58 ms

### Trail + Point
Using both optimizations and testing with rendering mode "Trails + Points".
Test using the _Main Asteroid Belt_ at different zoom levels within the solar system

**Before:**
-  42 - 75 fps
- 23.81 - 13.33 ms

**After:**
- 94 - 97 fps
- 10.64 - 10.31 ms

At most it reduced by 13.5 ms

##
### Test system
Intel Core 9 12900
64GB RAM
Nvidia RTX 3080 (10GB)
Screen resolution: 2560x1440