@GPayne 2024-05-31
Tested on linux. Everything looks great, and performance is definitely better than before.
I have a concern about the point size exponent. I was running with the main asteroid belt (the largest set) and accidentally clicked the exponent slider with too high a value, and was stuck with having to restart my laptop which had ground to a halt. It only has nvidia 1050 graphics, but other users with lesser GPUs might also encounter this problem.
I see the point of having the wide size exponent range to adjust based on the zoom distance. At least for this subject matter it should probably be capped at a lower value.