@Arohdin 2024-06-01
> I see the point of having the wide size exponent range to adjust based on the zoom distance. At least for this subject matter it should probably be capped at a lower value.

Yeah, it's an issue for both this renderable as well as the point cloud. Will think about this and try to fix it/work around it asap (at least for this renderable).